--- a/src/conversions/sample_rate.rs
+++ b/src/conversions/sample_rate.rs
@@ -53,7 +53,7 @@ where
         assert!(from >= 1);
         assert!(to >= 1);
 
-        // finding greatest common divisor
+        // finding the greatest common divisor
         let gcd = {
             #[inline]
             fn gcd(a: u32, b: u32) -> u32 {
@@ -259,12 +259,6 @@ mod test {
     use cpal::SampleRate;
     use quickcheck::quickcheck;
 
-    // TODO: Remove once cpal 0.12.2 is released and the dependency is updated
-    //  (cpal#483 implemented ops::Mul on SampleRate)
-    const fn multiply_rate(r: SampleRate, k: u32) -> SampleRate {
-        SampleRate(k * r.0)
-    }
-
     quickcheck! {
         /// Check that resampling an empty input produces no output.
         fn empty(from: u32, to: u32, n: u16) -> () {
@@ -296,7 +290,7 @@ mod test {
         ///   dropping a sample from each channel.
         fn divide_sample_rate(to: u32, k: u32, input: Vec<u16>, n: u16) -> () {
             let to = if to == 0 { return; } else { SampleRate(to) };
-            let from = multiply_rate(to, k);
+            let from = to * k;
             if k == 0 || n == 0 { return; }
 
             // Truncate the input, so it contains an integer number of frames.
@@ -320,7 +314,7 @@ mod test {
         ///  sample in the output matches exactly with the input.
         fn multiply_sample_rate(from: u32, k: u32, input: Vec<u16>, n: u16) -> () {
             let from = if from == 0 { return; } else { SampleRate(from) };
-            let to = multiply_rate(from, k);
+            let to = from * k;
             if k == 0 || n == 0 { return; }
 
             // Truncate the input, so it contains an integer number of frames.