1. `quickcheck`  1.x input values vary more than in 0.9 which finds overflow cases and excessive memory allocation problems in `Sample::lerp`.  Calculations now use f64 to avoid that. Conversion back to 16 bit resolution is done explicitly since `as` conversion may silently discard most significant bits. Alternatively, interpolation coefficient (`numerator/dividor`) may use e.g. `u16` instead of `u32`, or maybe even floating point. I believe `u32` precision is unnecessary there. But that requires updating sample rate converter which at the moment is the only user of this interpolation.
2. Documentation for `Sample::lerp` was incorrect. It said that calculations should follow `c * first + (c - 1) * second`. which gives `first` at `c==1` and `second` at `c==0` but actual implementations use the opposite `first` at `c==0` and `second` at `c==1`.
3. Fix documentation examples. Some of the examples are marked as `no_run` since they require audio devices and may actually make sounds while `cargo test` command is running.
4. Include documentation tests into CI. `--all-targets` switch excludes doc tests, so those are executed separately.
5. Non `experimental` builds exclude some code and in some code parts use mutually exclusive implementations, so I included those in CI too to keep experimental code compilable.