@PetrGlad 2024-11-16
Looking a other filters in `rodio` I am now curious if rodio should only provide basic functions and use https://github.com/RustAudio/dasp for complex processing instead. Some of the rodio functionality is already implemented there.

As I understand rodio in most cases can use generic parameters to avoid dyn references (but Sync and Mixer still use them). That probably can be slightly  more efficient. Any other reason rodio cannot use some existing external libraries for sound processing graph?