@dvdsk 2024-11-18
> I am now curious if rodio should only provide basic functions and use [RustAudio/dasp](https://github.com/RustAudio/dasp) for complex processing instead.

Rodio and dasp have fundamentally different goals, that makes it hard and maybe unwise to merge them or require users to be familiar with both. 
While rodio is an audio playback crate dasp is a signal processing suite. When Rodio is used its usually not the main feature of the application. The app might be a simple game, ui with sounds on click or podcast app. So Rodio should focus on doing most things efficient while getting out of the way of the user. That is why for example `Sink` exists. You do not need it, it adds nothing that can not be done with the other parts of rodio. The `Sink` API does however cover most use-cases and is easy to understand/use. (Its also far from perfect and has its own issues which we are going to address).

> Some of the rodio functionality is already implemented there.

Rodio predates dasp, so its easy to turn this around and ask, why do they not depend on rodio. Again its a difference in goals, dasp wants to have zero dependencies, rodio just wants to be easy to use (so no C-deps if possible).
 
> use [RustAudio/dasp](https://github.com/RustAudio/dasp) for complex processing instead

I also think rodio has most features dasp has since #602 landed. I might be mistaken there. And a Source that makes it easy to use dasp from rodio might be interesting.

The sinc interpolator in dasp is interesting, rodio needs an option for a slower but more hifi interpolator. See  https://github.com/RustAudio/rodio/issues/584.