@dvdsk 2024-11-18
Still have not found the time for the benchmark, as soon as I've done that ill post the results here and we can see what we do based on them.