@dvdsk 2024-11-29
tldr: Same kind of perf regression as before. Pretty strange, I would expect perf to be the same now I'll hunt for the cause tomorrow.

this pr:
```
Timer precision: 20 ns
resampler         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ no_resampling  2.399 ms      │ 2.795 ms      │ 2.494 ms      │ 2.505 ms      │ 100     │ 100
╰─ resample_to                  │               │               │               │         │
   ├─ 8000        2.137 ms      │ 2.385 ms      │ 2.187 ms      │ 2.197 ms      │ 100     │ 100
   ├─ 11025       2.377 ms      │ 2.436 ms      │ 2.392 ms      │ 2.395 ms      │ 100     │ 100
   ├─ 16000       3.085 ms      │ 3.173 ms      │ 3.12 ms       │ 3.122 ms      │ 100     │ 100
   ├─ 22050       3.501 ms      │ 5.609 ms      │ 3.52 ms       │ 3.543 ms      │ 100     │ 100
   ├─ 44100       2.432 ms      │ 3.042 ms      │ 2.557 ms      │ 2.585 ms      │ 100     │ 100
   ├─ 48000       6.456 ms      │ 6.505 ms      │ 6.476 ms      │ 6.477 ms      │ 100     │ 100
   ├─ 88200       10.71 ms      │ 10.82 ms      │ 10.75 ms      │ 10.75 ms      │ 100     │ 100
   ├─ 96000       12.14 ms      │ 14.13 ms      │ 12.18 ms      │ 12.21 ms      │ 100     │ 100
   ├─ 176400      20.01 ms      │ 23.26 ms      │ 20.07 ms      │ 20.12 ms      │ 100     │ 100
   ├─ 192000      22.95 ms      │ 26.81 ms      │ 23.02 ms      │ 23.16 ms      │ 100     │ 100
   ├─ 352800      40.27 ms      │ 40.79 ms      │ 40.4 ms       │ 40.4 ms       │ 100     │ 100
   ╰─ 384000      45.8 ms       │ 46.39 ms      │ 45.94 ms      │ 45.95 ms      │ 100     │ 100
```

main:
```
Timer precision: 20 ns
resampler         fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ no_resampling  2.283 ms      │ 2.75 ms       │ 2.392 ms      │ 2.426 ms      │ 100     │ 100
╰─ resample_to                  │               │               │               │         │
   ├─ 8000        2.342 ms      │ 2.423 ms      │ 2.356 ms      │ 2.36 ms       │ 100     │ 100
   ├─ 11025       2.518 ms      │ 2.88 ms       │ 2.533 ms      │ 2.54 ms       │ 100     │ 100
   ├─ 16000       3.207 ms      │ 3.283 ms      │ 3.25 ms       │ 3.248 ms      │ 100     │ 100
   ├─ 22050       3.599 ms      │ 3.635 ms      │ 3.613 ms      │ 3.614 ms      │ 100     │ 100
   ├─ 44100       2.326 ms      │ 3.04 ms       │ 2.48 ms       │ 2.51 ms       │ 100     │ 100
   ├─ 48000       6.368 ms      │ 6.407 ms      │ 6.384 ms      │ 6.384 ms      │ 100     │ 100
   ├─ 88200       9.995 ms      │ 10.83 ms      │ 10.28 ms      │ 10.28 ms      │ 100     │ 100
   ├─ 96000       11.92 ms      │ 12 ms         │ 11.95 ms      │ 11.95 ms      │ 100     │ 100
   ├─ 176400      19.33 ms      │ 36.44 ms      │ 19.4 ms       │ 19.57 ms      │ 100     │ 100
   ├─ 192000      22.5 ms       │ 22.71 ms      │ 22.57 ms      │ 22.57 ms      │ 100     │ 100
   ├─ 352800      39.32 ms      │ 54.21 ms      │ 39.43 ms      │ 39.67 ms      │ 100     │ 100
   ╰─ 384000      44.85 ms      │ 53.31 ms      │ 44.98 ms      │ 45.07 ms      │ 100     │ 100
```